requirements.txt
dhanhq
requests
python-dotenv
google-genai
flask
pyotp
PyJWT
tradingview-ta
